feat: add 'prepareGetData' method to 'LegacySamSelectionExtension' (#4)
* refactor: rename `CardCertificateComputationData` to `LegacyCardCertificateComputationData`

The class `CardCertificateComputationData` is renamed to `LegacyCardCertificateComputationData` across multiple files. The references and usage have been updated accordingly, including documentation and method signatures. This change is reflected in version `0.6.0` of the Keypop Calypso Crypto Legacy SAM API.
A new method named 'prepareGetData' was added to the 'LegacySamSelectionExtension' class, which schedules the execution of a "Get Data" command for a given tag.